# internal/swagger/swagger.yaml
basePath: /api
definitions:
...
  github_com_neu-software-practice_software-practice-backend_internal_dto.DiagnoseRequest:
    properties:
      cure:
        type: string
      diagnosis:
        type: string
      finish:
        type: boolean
    required:
    - diagnosis
    type: object
  github_com_neu-software-practice_software-practice-backend_internal_dto.DrugRequest:
    properties:
      drug_code:
        type: string
      drug_dosage:
        type: string
      drug_format:
        type: string
      drug_name:
        type: string
      drug_price:
        minimum: 0
        type: number
      drug_stock:
        minimum: 0
        type: integer
      drug_type:
        type: string
      drug_unit:
        type: string
      manufacturer:
        type: string
      mnemonic_code:
        type: string
    required:
    - drug_code
    - drug_name
    type: object
...
  github_com_neu-software-practice_software-practice-backend_internal_dto.MedicalRecordRequest:
    properties:
      allergy:
        type: string
      careful:
        type: string
      disease_ids:
        items:
          type: integer
        type: array
      history:
        type: string
      physique:
        type: string
      present:
        type: string
      present_treat:
        type: string
      proposal:
        type: string
      readme:
        type: string
    type: object
  github_com_neu-software-practice_software-practice-backend_internal_dto.PrescriptionItemInput:
    properties:
      drug_id:
        type: integer
      drug_number:
        minimum: 1
        type: integer
      drug_usage:
        type: string
    required:
    - drug_id
    - drug_number
    type: object
  github_com_neu-software-practice_software-practice-backend_internal_dto.PrescriptionRequest:
    properties:
      items:
        items:
          $ref: '#/definitions/github_com_neu-software-practice_software-practice-backend_internal_dto.PrescriptionItemInput'
        minItems: 1
        type: array
    required:
    - items
    type: object
...
  github_com_neu-software-practice_software-practice-backend_internal_dto.StockRequest:
    properties:
      delta:
        type: integer
    required:
    - delta
    type: object
  github_com_neu-software-practice_software-practice-backend_internal_pkg_response.Body:
    properties:
      data: {}
      error:
        $ref: '#/definitions/github_com_neu-software-practice_software-practice-backend_internal_pkg_response.ErrorBody'
      meta:
        $ref: '#/definitions/github_com_neu-software-practice_software-practice-backend_internal_pkg_response.Meta'
      success:
        type: boolean
    type: object
  github_com_neu-software-practice_software-practice-backend_internal_pkg_response.ErrorBody:
    properties:
      code:
        type: string
      message:
        type: string
    type: object
  github_com_neu-software-practice_software-practice-backend_internal_pkg_response.Meta:
    properties:
      limit:
        type: integer
      page:
        type: integer
      total:
        type: integer
    type: object
info:
  contact: {}
  description: 门诊全流程后端：挂号收费 / 门诊医生 / 检查 / 检验 / 处置 / 药房。
  title: 东软云医院 HIS 门诊管理 API
  version: "1.0"
paths:
...
  /pharmacy/drugs:
    post:
      consumes:
      - application/json
      parameters:
      - description: 药品信息
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_neu-software-practice_software-practice-backend_internal_dto.DrugRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_neu-software-practice_software-practice-backend_internal_pkg_response.Body'
      security:
      - BearerAuth: []
      summary: 药品入库/新增 (F5-3)
      tags:
      - pharmacy
  /pharmacy/drugs/{id}:
    delete:
      parameters:
      - description: 药品ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_neu-software-practice_software-practice-backend_internal_pkg_response.Body'
      security:
      - BearerAuth: []
      summary: 药品删除 (F5-3)
      tags:
      - pharmacy
    put:
      consumes:
      - application/json
      parameters:
      - description: 药品ID
        in: path
        name: id
        required: true
        type: integer
      - description: 药品信息
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_neu-software-practice_software-practice-backend_internal_dto.DrugRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_neu-software-practice_software-practice-backend_internal_pkg_response.Body'
      security:
      - BearerAuth: []
      summary: 药品信息维护 (F5-3)
      tags:
      - pharmacy
  /pharmacy/drugs/{id}/restock:
    post:
      consumes:
      - application/json
      parameters:
      - description: 药品ID
        in: path
        name: id
        required: true
        type: integer
      - description: 库存增量(正数入库)
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_neu-software-practice_software-practice-backend_internal_dto.StockRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_neu-software-practice_software-practice-backend_internal_pkg_response.Body'
      security:
      - BearerAuth: []
      summary: 库存调整/入库 (F5-3)
      tags:
      - pharmacy
  /pharmacy/prescriptions:
    get:
      parameters:
      - description: 病历号
        in: query
        name: case_number
        required: true
        type: string
      - description: 处方状态(默认 已缴费)
        in: query
        name: state
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_neu-software-practice_software-practice-backend_internal_pkg_response.Body'
      security:
      - BearerAuth: []
      summary: 待发药/处方查询 (F5-1)
      tags:
      - pharmacy
  /pharmacy/prescriptions/{id}/dispense:
    post:
      parameters:
      - description: 处方ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_neu-software-practice_software-practice-backend_internal_pkg_response.Body'
      security:
      - BearerAuth: []
      summary: 药房发药 (F5-1)
      tags:
      - pharmacy
  /pharmacy/prescriptions/{id}/refund:
    post:
      parameters:
      - description: 处方ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_neu-software-practice_software-practice-backend_internal_pkg_response.Body'
      security:
      - BearerAuth: []
      summary: 药房退药 (F5-2)
      tags:
      - pharmacy
  /pharmacy/transactions:
    get:
      parameters:
      - description: 病历号
        in: query
        name: case_number
        type: string
      - description: 动作(发药/退药)
        in: query
        name: action
        type: string
      - description: 页码
        in: query
        name: page
        type: integer
      - description: 每页条数
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_neu-software-practice_software-practice-backend_internal_pkg_response.Body'
      security:
      - BearerAuth: []
      summary: 药品交易记录 (F5-4)
      tags:
      - pharmacy
...
  /physician/history:
    get:
      parameters:
      - description: 病历号
        in: query
        name: case_number
        type: string
      - description: 姓名
        in: query
        name: name
        type: string
      - description: 页码
        in: query
        name: page
        type: integer
      - description: 每页条数
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_neu-software-practice_software-practice-backend_internal_pkg_response.Body'
      security:
      - BearerAuth: []
      summary: 看诊记录 (F2-5)
      tags:
      - physician
  /physician/patients:
    get:
      parameters:
      - description: 病历号
        in: query
        name: case_number
        type: string
      - description: 姓名
        in: query
        name: name
        type: string
      - description: 看诊状态
        in: query
        name: state
        type: integer
      - description: 页码
        in: query
        name: page
        type: integer
      - description: 每页条数
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_neu-software-practice_software-practice-backend_internal_pkg_response.Body'
      security:
      - BearerAuth: []
      summary: 患者查看 (F2-1)
      tags:
      - physician
  /physician/patients/counts:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_neu-software-practice_software-practice-backend_internal_pkg_response.Body'
      security:
      - BearerAuth: []
      summary: 患者统计 (排队/已看诊, F2-1)
      tags:
      - physician
  /physician/registers/{id}/consult:
    post:
      parameters:
      - description: 挂号ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_neu-software-practice_software-practice-backend_internal_pkg_response.Body'
      security:
      - BearerAuth: []
      summary: 创建病历开始看诊 (F2-1)
      tags:
      - physician
  /physician/registers/{id}/diagnosis:
    put:
      consumes:
      - application/json
      parameters:
      - description: 挂号ID
        in: path
        name: id
        required: true
        type: integer
      - description: 诊断结果
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_neu-software-practice_software-practice-backend_internal_dto.DiagnoseRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_neu-software-practice_software-practice-backend_internal_pkg_response.Body'
      security:
      - BearerAuth: []
      summary: 门诊确诊 (F2-8)
      tags:
      - physician
  /physician/registers/{id}/medical-record:
    get:
      parameters:
      - description: 挂号ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_neu-software-practice_software-practice-backend_internal_pkg_response.Body'
      security:
      - BearerAuth: []
      summary: 读取病历首页 (F2-2)
      tags:
      - physician
    put:
      consumes:
      - application/json
      parameters:
      - description: 挂号ID
        in: path
        name: id
        required: true
        type: integer
      - description: 病历内容
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_neu-software-practice_software-practice-backend_internal_dto.MedicalRecordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_neu-software-practice_software-practice-backend_internal_pkg_response.Body'
      security:
      - BearerAuth: []
      summary: 保存病历首页 (F2-2)
      tags:
      - physician
  /physician/registers/{id}/prescriptions:
    post:
      consumes:
      - application/json
      parameters:
      - description: 挂号ID
        in: path
        name: id
        required: true
        type: integer
      - description: 处方明细
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_neu-software-practice_software-practice-backend_internal_dto.PrescriptionRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_neu-software-practice_software-practice-backend_internal_pkg_response.Body'
      security:
      - BearerAuth: []
      summary: 开立处方 (F2-9)
      tags:
      - physician
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"